Add files via upload
i was asked to achieve the following
 in task1: Ask the user to enter a message then display it using the different 
html heading tags (from <h1> to <h6>) using Loops. DO NOT write the 
header element explicitly in your script! 

 in task2: Write a script that takes from the user n values and returns their sum, 
stop receiving values from user when he enters 0 or sum exceeds 100, 
check that the entered data is numeric and inform the user with the total 
sum of the entered values in console.

in task3:Ask the user to enter three numerical values x, y and z. The function 
should check if x is divisible by y only or z only or both y and z. 
Example: If user entered values are 10, 2, and 5 then the output will be: 
10 is divisible by both 2 and 5. 
If user entered values are 10, 2, and 4 then the output will be: 10 is 
divisible by 2 only. 
If user entered values are 10, 5, and 4 then the output will be: 10 is 
divisible by 5 only. 

in task4:Ask the user to enter 3 values where x and y are two numeric values 
while z is a string value. The function should display the following: 
• The range of numbers between x and y depending on the value of z 
according to the following: 
o If z=”odd” the function will display only the odd values between x 
and y. x and/or y are included if any is odd. 
o If z=”even” the function will display only the even values between x 
and y. x and/or y are included if any is even. 
o If z=”no” the function will display all values between x and y. x and y 
are included. 
o If z=”odd” the function will display only the odd values between x 
and y. 
o Make sure that the user entered the proper data type for each 
parameter, if not, inform the user to re-enter the required 
parameters. 
• The sum of the displayed numbers. 
• Apply a displaying style in console. 
The output will be either descending (if x>y) or ascending order (if x<y) 
Example: If user entered values are 9,15 and answered the question 
about the even and odd with “no” then the output will be: 
9,10,11,12,13,14,15  and their sum value that is 84 
If the answer is “odd” then the output will be: 9,11,13,15 and their sum 
value that is 48 
If the answer is “even” then the output will be: 10,12,14 and their sum 
value that is 36