Add files via upload
Revise the default functionality of the service worker, which initially only allowed it to receive requests from the main thread and directly send them to the server for responses. This approach led to slow website loading, relying solely on server-side rendering. To address this issue and enhance website speed, my solution involves modifying the service worker to accept requests from the main thread and enable responses to originate from either the server or cache storage, thereby optimizing the loading speed through a combination of server-side and client-side rendering